added log

+++ src/routes.js
@@ -6,6 +6,12 @@ import { isValid, exec, validateDetectors } from "./helper"
 const analyzeRouter = async function(req, res, next){
     const { disableDetectors, enableDetectors, source: { sources, target }, data } = req.body
 
+    const contract = sources[target].content;
+    const fileName = target.split('/').pop();
+    const fileDir = `${path.dirname(target)}`;
+    const filePath = fileDir + '/'+fileName;
+    const outputFile = `${path.dirname(target)}/output.json`
+
     let cmd = `slither ${filePath} --disable-solc-warnings --json ${outputFile}`
     
     let response = {
@@ -31,29 +37,25 @@ const analyzeRouter = async function(req, res, next){
         cmd = `${cmd} --exclude ${disableDetectors}`
     }
 
-    const contract = sources[target].content;
-    const fileName = target.split('/').pop();
-    const fileDir = `${path.dirname(target)}`;
-    shell.mkdir('-p', fileDir);
-    const filePath = fileDir + '/'+fileName;
-    const outputFile = `${path.dirname(target)}/output.json`
+    let unlinkOutput = false
 
     try {
-
+        shell.mkdir('-p', fileDir);
         fs.writeFileSync(filePath, contract);
         // execute slither command
         let {stderr} = await exec(cmd)
         response["output"] = stderr
 
     } catch(error) {
-
+        
         let data = JSON.parse(fs.readFileSync(outputFile, 'utf8'))
+        unlinkOutput = true
         response.error = data
 
     } finally {
         // delete file
         await fs.unlinkSync(filePath)
-        await fs.unlinkSync(outputFile)
+        if(unlinkOutput) await fs.unlinkSync(outputFile)
     }
 
     return res.status(200).json(response)
